pyproject.toml
[project]
name = "jx3-DrytronII"
version = "0.1.0"
description = "JX3 Seaech BOT"
readme = "README.md"
requires-python = "==3.11.*"
authors = [
    {name = "KolinLoa", email = "kolinloa@petalmail.com"},
]
dependencies = [
    "jx3api>=2024.10.21",
    "nonebot2[fastapi]>=2.3.3",
    "nonebot-adapter-onebot==2.4.5",
    "nb-cli>=1.4.2",
    "zhipuai>=2.1.5.20230904",
]
license = {text = "MIT"}

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = []
plugin_dirs = ["src/plugins"]
builtin_plugins = ["echo"]

[tool.pdm]
distribution = false
